Update our context for go 1.7
When building with go 1.7, use the standard library's context package.
For packages below 1.7, use the golang.org/x/net/context package.
